@@ -12,7 +12,7 @@ keywords = ["command-line", "task", "runner", "development", "utility"]
 license = "CC0-1.0"
 readme = "crates-io-readme.md"
 repository = "https://github.com/casey/just"
-rust-version = "1.63"
+rust-version = "1.70"
 
 [workspace]
 members = [".", "crates/*"]
