@@ -231,11 +231,8 @@ impl<'src> Lexer<'src> {
     // The width of the error site to highlight depends on the kind of error:
     let length = match kind {
       UnterminatedString | UnterminatedBacktick => {
-        let kind = match StringKind::from_token_start(self.lexeme()) {
-          Some(kind) => kind,
-          None => {
-            return self.internal_error("Lexer::error: expected string or backtick token start")
-          }
+        let Some(kind) = StringKind::from_token_start(self.lexeme()) else {
+          return self.internal_error("Lexer::error: expected string or backtick token start");
         };
         kind.delimiter().len()
       }
@@ -813,9 +810,7 @@ impl<'src> Lexer<'src> {
   /// Cooked string: "[^"]*" # also processes escape sequences
   /// Raw string:    '[^']*'
   fn lex_string(&mut self) -> CompileResult<'src> {
-    let kind = if let Some(kind) = StringKind::from_token_start(self.rest()) {
-      kind
-    } else {
+    let Some(kind) = StringKind::from_token_start(self.rest()) else {
       self.advance()?;
       return Err(self.internal_error("Lexer::lex_string: invalid string start"));
     };
